rename: markerIntanceStore → markerInstanceStore로 이름 변경
[#93]
