+++ frontend/src/stores/markerInstanceStore.ts
